Refactor utils
- 폴더로 묶음
- 필요 없어진 get 삭제
