+++ src/utils/utils.js
@@ -1,7 +1,4 @@
-export function get(key) {
-  return (obj) => obj[key];
-}
-
+// eslint-disable-next-line import/prefer-default-export
 export function equal(key, value) {
   return (obj) => obj[key] === value;
 }
