+++ src/utils/utils.test.js
@@ -1,0 +1,13 @@
+import { equal } from './utils';
+
+test('equal', () => {
+  const state = {
+    name: '홍길동',
+  };
+
+  const f = equal('name', '홍길동');
+  const g = equal('name', '임꺽정');
+
+  expect(f(state)).toBeTruthy();
+  expect(g(state)).toBeFalsy();
+});
